requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
plaid-python
python-dotenv
SQLAlchemy
psycopg2-binary
alembic
